refactor(Application):into one form
