+++ src/app/applications/applications.module.ts
@@ -16,6 +16,7 @@ import {ValidationApplicationComponent} from '../validation-application/validati
 import {AutocompleteLibModule} from 'angular-ng-autocomplete';
 import {ApplicationDetailComponent} from './application-detail/application-detail.component';
 import {SharedDirectivesModule} from '../shared/shared_modules/shared_directives.module';
+import { ApplicationFormularComponent } from './application-formular/application-formular.component';
 
 /**
  * Applications Module.
@@ -39,7 +40,8 @@ import {SharedDirectivesModule} from '../shared/shared_modules/shared_directives
               AddcloudapplicationComponent,
               TypeOverviewComponent,
               ValidationApplicationComponent,
-              ApplicationDetailComponent
+              ApplicationDetailComponent,
+              ApplicationFormularComponent
             ],
             exports: [ApplicationDetailComponent]
           })
